feat: Updates install script.

+++ install.sh
@@ -2,20 +2,11 @@
 
 shopt -s extglob
 
+THIS_PATH=$(dirname "${BASH_SOURCE[0]}")
 XDG_CONFIG_HOME=${XDG_CONFIG_HOME:-"$HOME/.config"}
 dry_run="0"
 
 configDir=${XDG_CONFIG_HOME}/nvim
-files=(
-	init.lua
-	.neoconf.json
-	stylua.toml
-)
-# folders=(
-# 	after
-# 	snippets
-# 	spell
-# )
 
 log() {
 	if [[ $dry_run == "1" ]]; then
@@ -37,19 +28,32 @@ while [[ $# -gt 0 ]]; do
 done
 
 mkdir -p "$configDir"
-# for file in "${files[@]}"; do
-for file in *; do
-	if [[ -f $file ]]; then
-		log "Copying file: $file to $configDir"
+
+pushd "$THIS_PATH" &>/dev/null || exit 1
+(
+	# Copy files
+	for file in *; do
+		if [[ -f $file ]] && [[ ! "$file" == "install.sh" ]]; then
+			log "Copying file: $file to $configDir"
+			if [[ $dry_run == "0" ]]; then
+				cp "$file" "$configDir"
+			fi
+		fi
+	done
+
+	if [[ -f .neoconf.json ]]; then
+		log "Copying file: .neoconf.json to $configDir"
 		if [[ $dry_run == "0" ]]; then
-			cp "$file" "$configDir"
+			cp .neoconf.json "$configDir"
 		fi
 	fi
-done
 
-# for folder in !(.)*/; do
-# 	log "Copying folder: $folder to $configDir"
-# 	if [[ $dry_run == "0" ]]; then
-# 		cp -R "$folder" "$configDir"
-# 	fi
-# done
+	for folder in !(.)*/; do
+		log "Copying folder: $folder to $configDir"
+		if [[ $dry_run == "0" ]]; then
+			cp -R "$folder" "$configDir"
+		fi
+	done
+)
+
+popd &>/dev/null || exit 1
